+++ src/context/input_reader.rs
@@ -299,7 +299,7 @@ impl InputReader<'_, '_> {
 /// use bevy_enhanced_input::prelude::*;
 ///
 /// # let mut app = App::new();
-/// app.add_systems(PreUpdate, disable_mouse.before(EnhancedInputSet::Update));
+/// app.add_systems(PreUpdate, disable_mouse.before(EnhancedInputSystems::Update));
 ///
 /// fn disable_mouse(
 ///     mut action_sources: ResMut<ActionSources>,