+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.19.2] - 2025-10-30
+
 ### Added
 
 - `Combo` input condition.
@@ -468,7 +470,8 @@ This update features a big rewrite into a component-based API. The core concepts
 
 Initial release.
 
-[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.19.1...HEAD
+[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.19.2...HEAD
+[0.19.2]: https://github.com/simgine/bevy_replicon/compare/v0.19.1...v0.19.2
 [0.19.1]: https://github.com/simgine/bevy_replicon/compare/v0.19.0...v0.19.1
 [0.19.0]: https://github.com/simgine/bevy_replicon/compare/v0.18.2...v0.19.0
 [0.18.2]: https://github.com/simgine/bevy_replicon/compare/v0.18.1...v0.18.2